my_civil_service_project/pyproject.toml
[tool.poetry]
name = "civil-service"
version = "0.1.0"
description = ""
authors = ["yukendh"]
readme = "README.md"
packages = [{include = "civil_service"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "3.2.17"
djangorestframework = "3.14.0"
django-rest-framework-social-oauth2 = "1.0.4"
psycopg2-binary = "2.9.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
